Gemini + Faiss/requirements.txt
fastapi
uvicorn
langchain
langchain-community
langchainhub
langchain-openai
sentence-transformers
gTTS
deep-translator
python-dotenv